requirements.txt
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.0
asgiref==3.8.1
blinker==1.9.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
clerk-backend-api==5.0.2
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
Django==5.2.1
dnspython==2.8.0
docstring_parser==0.17.0
et_xmlfile==2.0.0
Flask==3.1.3
fpdf2==2.8.7
O365==2.0.37
flask-cors==6.0.2
geopandas==1.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
numpy==2.3.0
openai==2.14.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
pyodbc==5.2.0
pyogrio==0.11.0
pypdfium2==5.2.0
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2025.2
pywin32==310
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.3
shapely==2.1.1
simplekml==1.3.6
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.6
xlrd==2.0.2
xlwings==0.33.15
xlwt==1.3.0
